refactor(coin-module-boilerplate): flatten logic/ folder per ADR-049
Drop the nested account/history/transaction subfolders in favor of a
flat logic/ (one file per CoinModuleApi method), consolidate its tests
into logic/tests/, and remove the logic/ barrel along with the dead
public ./logic package export. Cross-layer helpers shared with bridge/
move to common-logic/.

LIVE-35517
