+++ libs/coin-modules/coin-module-boilerplate/src/common-logic/utils.ts
