--- libs/coin-modules/coin-module-boilerplate/src/logic/account/getNextSequence.ts
@@ -1,6 +1,0 @@
-import { getNextSequence } from "../../network/node";
-
-// Could be getAccountInfo so it is used in both bridge and api
-export async function getNextValidSequence(address: string): Promise<number> {
-  return await getNextSequence(address);
-}
