+++ libs/coin-modules/coin-module-boilerplate/src/logic/broadcast.ts
@@ -1,4 +1,4 @@
-import { submit } from "../../network/node";
+import { submit } from "../network/node";
 
 export async function broadcast(signedTx: string): Promise<string> {
   const submittedPayment = await submit(signedTx);
