+++ libs/coin-modules/coin-module-boilerplate/src/logic/combine.ts
@@ -1,4 +1,4 @@
-import { encode } from "../utils";
+import { encode } from "../common-logic/utils";
 
 // Combines signature with raw transaction
 export function combine(transaction: string, signature: string, publicKey?: string): string {
