+++ libs/coin-modules/coin-module-boilerplate/src/logic/getBalance.ts
@@ -1,5 +1,5 @@
 import { Balance } from "@ledgerhq/coin-module-framework/api/types";
-import { getAccountInfo } from "../../network/node";
+import { getAccountInfo } from "../network/node";
 
 // Could be getAccountInfo so it is used in both bridge and api
 export async function getBalance(address: string): Promise<Balance[]> {
