--- libs/coin-modules/coin-module-boilerplate/src/logic/index.ts
@@ -1,11 +1,0 @@
-export { broadcast } from "./transaction/broadcast";
-export { combine } from "./transaction/combine";
-export { craftTransaction } from "./transaction/craftTransaction";
-export { estimateFees } from "./transaction/estimateFees";
-export { getBalance } from "./account/getBalance";
-export { lastBlock } from "./history/lastBlock";
-export { listOperations } from "./history/listOperations";
-export { isRecipientValid } from "./utils";
-export { getNextValidSequence } from "./account/getNextSequence";
-
-export { parseAPIValue } from "./common";
